Refactor compiler into separate files
