+++ compiler/index.ts
@@ -1,0 +1,9 @@
+export {
+  RoutineType,
+  CompilerOptions,
+  CompileResult,
+  LineCompiler,
+  compileInstruction,
+  compileProgram,
+} from './base';
+export { CompiledCmd, CommandCompiler, compileCommands } from './command';
